docker/requirements.txt
numpy
Pillow
tensorflow
keras
flask